# src/main/resources/swagger_config/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8081
  description: Generated server url
paths:
  /api/v1/room_facility:
    put:
      tags:
      - Room Facility
      operationId: updateRoomFacility
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomFacilityRequest'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseRoomFacilityResponse'
    post:
      tags:
      - Room Facility
      operationId: createRoomFacility
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomFacilityRequest'
        required: true
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseRoomFacilityResponse'
...
  /api/v1/room_facility/{id}:
    get:
      tags:
      - Room Facility
      operationId: getRoomFacilityById
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseRoomFacilityResponse'
    delete:
      tags:
      - Room Facility
      operationId: deleteRoomFacility
      parameters:
      - name: id
        in: path
        required: true
        style: simple
        explode: false
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CommonResponseString'
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/inline_response_400'
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommonResponseString'
...
components:
  schemas:
    CommonResponseString:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          type: string
    CommonResponseErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        status:
          type: integer
          format: int32
        timestamp:
          type: integer
          format: int64
    RoomFacilityRequest:
      type: object
      properties:
        roomFacilityId:
          type: string
        roomFacilityName:
          type: string
    CommonResponseRoomFacilityResponse:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        message:
          type: string
        data:
          $ref: '#/components/schemas/RoomFacilityResponse'
    RoomFacilityResponse:
      type: object
      properties:
        roomId:
          type: string
        roomFacilityName:
          type: string
...
    inline_response_400:
      oneOf:
      - $ref: '#/components/schemas/CommonResponseErrorResponse'
      - $ref: '#/components/schemas/CommonResponseString'